+++ includes/unicode.h
@@ -9,8 +9,8 @@
 #ifndef _UNICODE_H_
 #define _UNICODE_H_
 
-uint32_t get_number_encode_utf8(int value);
-uint32_t get_number_decode_utf8(uint8_t byte);
+uint64_t get_number_encode_utf8(int value);
+uint64_t get_number_decode_utf8(uint8_t byte);
 uint8_t encode_utf8(uint8_t *buff, int value);
 int decode_utf8(const uint8_t *byte_ptr, size_t length);
